move & fix SanitizeKubernetesName helper
