+++ pkg/common/helper_test.go
@@ -23,7 +23,6 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
-	"strings"
 	"testing"
 	"time"
 
@@ -898,103 +897,6 @@ func (suite *StripImageTagTestSuite) TestStripImageTag() {
 	}
 }
 
-type SanitizeKubernetesNameTestSuite struct {
-	suite.Suite
-}
-
-func (suite *SanitizeKubernetesNameTestSuite) TestSanitizeKubernetesName() {
-	for _, testCase := range []struct {
-		name            string
-		prefix          string
-		value           string
-		forGenerateName bool
-		expected        string
-	}{
-		{
-			name:     "LowercasesAndReplacesInvalidChars",
-			value:    "Registry.Example.Com/My_Func:v1.2",
-			expected: "registry-example-com-my-func-v1-2",
-		},
-		{
-			name:     "RunOfInvalidCharsBecomesOneDash",
-			value:    "a...b",
-			expected: "a-b",
-		},
-		{
-			name:     "LeadingAndTrailingDashesTrimmed",
-			value:    "--a-b--",
-			expected: "a-b",
-		},
-		{
-			name:     "PrefixUsedVerbatim",
-			prefix:   "registry-login-azure-",
-			value:    "myregistry.azurecr.io",
-			expected: "registry-login-azure-myregistry-azurecr-io",
-		},
-		{
-			name:     "EmptyValueLeavesNoDanglingDash",
-			prefix:   "registry-login-aws-",
-			value:    "///:::",
-			expected: "registry-login-aws",
-		},
-		{
-			name:     "TruncatedToLabelLimit",
-			prefix:   "registry-login-aws-",
-			value:    strings.Repeat("a", 80),
-			expected: "registry-login-aws-" + strings.Repeat("a", KubernetesDomainLevelMaxLength-len("registry-login-aws-")),
-		},
-		{
-			name:            "GenerateNameAppendsDashAndReservesSuffix",
-			prefix:          "nuclio-buildjob-",
-			value:           "my-func:latest",
-			forGenerateName: true,
-			expected:        "nuclio-buildjob-my-func-latest-",
-		},
-		{
-			name:            "GenerateNameTruncatesLeavingRoomForSuffix",
-			prefix:          "nuclio-buildjob-",
-			value:           strings.Repeat("x", 80),
-			forGenerateName: true,
-			expected:        "nuclio-buildjob-" + strings.Repeat("x", 41) + "-",
-		},
-		{
-			name:            "GenerateNameTruncationLandingOnDashIsTrimmed",
-			prefix:          "nuclio-buildjob-",
-			value:           strings.Repeat("a", 41) + "." + strings.Repeat("b", 40),
-			forGenerateName: true,
-			expected:        "nuclio-buildjob-" + strings.Repeat("a", 41) + "-",
-		},
-	} {
-		suite.Run(testCase.name, func() {
-			result, err := SanitizeKubernetesName(testCase.prefix, testCase.value, testCase.forGenerateName)
-
-			suite.Require().NoError(err)
-			suite.Equal(testCase.expected, result)
-
-			// the result must always fit a Kubernetes name, including the generated random suffix
-			totalLength := len(result)
-			if testCase.forGenerateName {
-				totalLength += generateNameRandomSuffixLength
-			}
-			suite.LessOrEqual(totalLength, KubernetesDomainLevelMaxLength)
-		})
-	}
-}
-
-func (suite *SanitizeKubernetesNameTestSuite) TestSanitizeKubernetesNameErrorsOnInvalidPrefix() {
-	for _, prefix := range []string{"Nuclio-", "nuclio_", "nuclio/", "nuclio ", "has.dot-"} {
-		_, err := SanitizeKubernetesName(prefix, "value", false)
-		suite.Require().Error(err, "prefix %q should be rejected", prefix)
-	}
-}
-
-func (suite *SanitizeKubernetesNameTestSuite) TestSanitizeKubernetesNameAcceptsValidPrefix() {
-	for _, prefix := range []string{"", "nuclio-", "a-b-c-", "abc123"} {
-		_, err := SanitizeKubernetesName(prefix, "value", false)
-		suite.Require().NoError(err, "prefix %q should be accepted", prefix)
-	}
-}
-
 func TestHelperTestSuite(t *testing.T) {
 	suite.Run(t, new(RetryUntilSuccessfulTestSuite))
 	suite.Run(t, new(RetryUntilSuccessfulOnErrorPatternsTestSuite))
@@ -1009,5 +911,4 @@ func TestHelperTestSuite(t *testing.T) {
 	suite.Run(t, new(ContainsPathTraversalTestSuite))
 	suite.Run(t, new(EnvWithLegacyKeyTestSuite))
 	suite.Run(t, new(StripImageTagTestSuite))
-	suite.Run(t, new(SanitizeKubernetesNameTestSuite))
 }
